+++ package/thunderstore/HipLantern/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.23
+* custom slots for AzuEPI and ExtraSlots can now be dynamically enabled and disabled (this fixes the issue when slot is disabled on a client and enabled on a server)
+* utility and custom type value can now also be changed without restart
+
 # 1.0.22
 * minor improvements